backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23  # For async support
aiosqlite==0.19.0           # Async SQLite driver

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Data Validation & Serialization
pydantic==2.5.0
email-validator==2.1.0

# External APIs
google-generativeai==0.3.0  # Gemini AI
google-maps-services-python==0.4.0  # Google Maps API

# Development & Testing
httpx==0.25.2               # HTTP client for testing
pytest==7.4.3               # Testing framework
pytest-asyncio==0.21.1     # Async testing support

# Environment & Configuration
python-dotenv==1.0.0        # Environment variables

# Date/Time handling
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10              # Fast JSON library

# Utilities
typing-extensions==4.8.0    # Type hints